Refactor Junit5 for BaseTestSHA3_224KAT and BaseTestSHA3_256KAT tests
This change consolidates the tests previously associated with BaseTestSHA3_224KAT into the parameterized TestSHA3_224KAT test class, and the tests from BaseTestSHA3_256KAT into the parameterized TestSHA3_256KAT test class.
